Update object position on touch

+++ src/com/app/ui/assignment/barrelrace/objects/Horse.java
@@ -15,7 +15,7 @@ public class Horse {
         mPaint.setColor(Color.RED);
     }
     
-    public void draw(int x, int y, int radius, Canvas c) {
+    public void draw(float x, float y, int radius, Canvas c) {
         c.drawCircle(x, y, radius, mPaint);
     }
 }
